feat: use route groups instead of seperate pages
