+++ src/app/(view)/layout.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export default function ViewLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <div className="bg-[url(/asset/background.jpg)] bg-fixed bg-cover bg-no-repeat bg-center" tabIndex={-1}>
+      {children}
+    </div>
+  );
+}
